Updates to examples directory to prepare for better examples soon
