--- EXAMPLE/pages/00_example_pages/casual_dresses_page.yml
@@ -1,3 +1,0 @@
-EXPECTED_DATA:
-  DEFAULT:
-    :title_static_text: "CASUAL DRESSES \nThere is 1 product."
