--- EXAMPLE/pages/00_example_pages/dresses_page.yml
@@ -1,3 +1,0 @@
-EXPECTED_DATA:
-  DEFAULT:
-    :title_static_text: "DRESSES \nThere are 5 products."
