+++ EXAMPLE/pages/01_catalog_pages/casual_dresses_page.yml
@@ -1,0 +1,3 @@
+EXPECTED_DATA:
+  DEFAULT:
+    :title_static_text: "CASUAL DRESSES \nThere is 1 product."
