+++ EXAMPLE/pages/01_catalog_pages/dresses_page.yml
@@ -1,0 +1,3 @@
+EXPECTED_DATA:
+  DEFAULT:
+    :title_static_text: "DRESSES \nThere are 5 products."
