+++ EXAMPLE/pages/example_storefront_root_page.rb
@@ -1,0 +1,10 @@
+
+
+class ExampleStorefrontRootPage < CorePage
+
+  def begin_new_session
+    browser.goto(@world.configuration['ENVIRONMENT']["URL"])
+    @world.assert_and_set_page(HomePage)
+  end
+
+end
