+++ EXAMPLE/step_definitions/hooks.rb
@@ -1,5 +1,5 @@
 
 Before do |scenario|
-  set_root_page(ExampleRootPage)
+  set_root_page(ExampleStorefrontRootPage)
   @browser_engine.create_new_browser
 end
